now with scripts that should work no matter what

+++ app/testing/live/run.sh
@@ -2,7 +2,6 @@
 
 set -euo pipefail
 
-# error trap
 trap 'echo "[ERROR] Command failed at line $LINENO: $BASH_COMMAND" >&2' ERR
 
 IMAGE_NAME=live-poc
@@ -17,24 +16,26 @@ if [ "$(sudo docker ps -aq -f name=$CONTAINER_NAME)" ]; then
     echo "Removing existing container..."
     sudo docker stop $CONTAINER_NAME
     echo "Successfully stopped the container"
-    # sudo docker rm $CONTAINER_NAME
+    sudo docker rm $CONTAINER_NAME || true
 fi
 
 echo "Removing all recording and live feed background processes"
-pids=$(ps -eo pid,cmd | grep "[b]ash scripts/" | awk '{print $1}')
+pids=$(ps -eo pid,cmd | grep "[b]ash scripts/" | awk '{print $1}' || true)
 
-for pid in $pids; do
-    echo "Found script PID: $pid"
-    
-    # Print full command used to launch the script
-    cmd=$(ps -p "$pid" -o cmd=) || true
-    echo "The command was: $cmd"
-    
-    # Kill the script itself
-    kill "$pid"
+if [[ -n "$pids" ]]; then
+    for pid in $pids; do
+        echo "Found script PID: $pid"
+        
+        # Print full command used to launch the script
+        cmd=$(ps -p "$pid" -o cmd=) || true
+        echo "The command was: $cmd"
+        
+        # Kill the script itself
+        kill "$pid"
 
-    echo "" 
-done
+        echo "" 
+    done
+fi
 
 # ps aux | grep [f]fmpeg
 pkill -x ffmpeg || true
